Move power ups to their own files
