+++ Scripts/Exceptions/NodeNotFoundException.cs
@@ -5,5 +5,7 @@ namespace PhotonPhighters.Scripts.Exceptions;
 public class NodeNotFoundException : Exception
 {
   public NodeNotFoundException(string message)
-    : base(message) { }
+    : base(message)
+  {
+  }
 }
