+++ Scripts/GoSharper/GsTimerFactory.cs
@@ -42,12 +42,7 @@ public static class GsTimerFactory
   /// </returns>
   public static Timer OneShotStartedTimer(double waitTime, Action onTimeout)
   {
-    var timer = new Timer
-    {
-      Autostart = true,
-      OneShot = true,
-      WaitTime = waitTime
-    };
+    var timer = new Timer { Autostart = true, OneShot = true, WaitTime = waitTime };
 
     if (onTimeout != null)
     {
@@ -68,12 +63,7 @@ public static class GsTimerFactory
   /// </returns>
   public static Timer StartedTimer(int waitTime)
   {
-    var timer = new Timer
-    {
-      Autostart = true,
-      OneShot = false,
-      WaitTime = waitTime
-    };
+    var timer = new Timer { Autostart = true, OneShot = false, WaitTime = waitTime };
 
     return timer;
   }
