+++ Scripts/GoSharper/Instancing/InstantiableAttribute.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Godot;
 
 namespace PhotonPhighters.Scripts.GoSharper.Instancing;
