+++ Scripts/PowerUps/PowerUpAppliers/GeneratorEngine.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the.NET Foundation under one or more agreements.
+// The.NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace PhotonPhighters.Scripts.PowerUps.PowerUpAppliers;
+
+public static partial class PowerUps
+{
+  public class GeneratorEngine : Scripts.PowerUps.PowerUps.AbstractPowerUpApplier
+  {
+    public override string Name => "Generator Engine";
+
+    public override Scripts.PowerUps.PowerUps.Rarity Rarity => Scripts.PowerUps.PowerUps.Rarity.Rare;
+
+    protected override void _Apply(Player playerWhoSelected, Player otherPlayer)
+    {
+      playerWhoSelected.Gun.FireRate += 0.9f;
+      playerWhoSelected.Gun.BulletSpread *= 1.06f;
+    }
+  }
+}
