+++ Scripts/PowerUps/PowerUpAppliers/PhotonEnlarger.cs
@@ -1,0 +1,25 @@
+﻿// Licensed to the.NET Foundation under one or more agreements.
+// The.NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Godot;
+
+namespace PhotonPhighters.Scripts.PowerUps.PowerUpAppliers;
+
+public static partial class PowerUps
+{
+  public class PhotonEnlarger : Scripts.PowerUps.PowerUps.AbstractPowerUpApplier
+  {
+    public override string Name => "Photon Enlarger";
+
+    public override Scripts.PowerUps.PowerUps.Rarity Rarity => Scripts.PowerUps.PowerUps.Rarity.Common;
+
+    protected override void _Apply(Player playerWhoSelected, Player otherPlayer)
+    {
+      playerWhoSelected.Gun.BulletSizeFactor += 1.5f;
+      playerWhoSelected.Gun.BulletDamage = Mathf.RoundToInt(playerWhoSelected.Gun.BulletDamage * 1.333f);
+      playerWhoSelected.Gun.BulletSpeed -= 150.0f;
+      playerWhoSelected.Gun.BulletSpread *= 1.25f;
+    }
+  }
+}
