+++ Scripts/PowerUps/PowerUpAppliers/SteelBootsCurse.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the.NET Foundation under one or more agreements.
+// The.NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace PhotonPhighters.Scripts.PowerUps.PowerUpAppliers;
+
+public static partial class PowerUps
+{
+  public class SteelBootsCurse : Scripts.PowerUps.PowerUps.AbstractPowerUpApplier
+  {
+    public override string Name => "Steel Boots Curse";
+    public override Scripts.PowerUps.PowerUps.Rarity Rarity => Scripts.PowerUps.PowerUps.Rarity.Rare;
+
+    protected override void _Apply(Player playerWhoSelected, Player otherPlayer)
+    {
+      otherPlayer.PlayerMovementDelegate.JumpForce /= 1.33f;
+    }
+  }
+}
